Move create_cache to storage submodule
