+++ src/tristan/command_line/images/gated_images_cli.py
@@ -18,11 +18,7 @@ from hdf5plugin import Bitshuffle
 from nexgen.nxs_copy import copy_tristan_nexus
 
 from ... import WithLocalDistributedCluster, compute_with_progress
-from ...binning import (
-    create_cache,
-    event_block_to_image_cache,
-    find_preceding_bin_edge_index,
-)
+from ...binning import event_block_to_image_cache, find_preceding_bin_edge_index
 from ...data import (
     cue_keys,
     cue_times,
@@ -37,6 +33,7 @@ from ...data import (
     time_bin_key,
     valid_events,
 )
+from ...storage import create_cache
 from .. import check_output_file, data_files, triggers
 from . import determine_image_size
